Move Bolt\Menu\AdminMenuBuilder to Bolt\Menu\Builder\AdminMenu
